Fixed a warning "comparison of integer expressions of different signedness"

+++ src/raft_server.cxx
@@ -1128,7 +1128,7 @@ bool raft_server::check_leadership_validity() {
              min_quorum_size);
 
         const auto nr_peers_list = get_not_responding_peers();
-        assert(nr_peers_list.size() == nr_peers);
+        assert(nr_peers_list.size() == static_cast<std::size_t>(nr_peers));
         for (auto& peer : nr_peers_list) {
             if (peer->is_lost()) {
                 continue;
